refactor: import taxonomy view
